Move rheed image results method
